feat(auth): extend authentication and authorization mechanisms across controllers

+++ src/modules/users/users.controller.ts
@@ -6,10 +6,15 @@ import {
     Patch,
     Param,
     Delete,
+    UseGuards,
+    Req,
+    UnauthorizedException,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
+import { AccessTokenGuard } from '../auth/guards/auth.guard';
+import type { Request } from 'express';
 
 @Controller('users')
 export class UsersController {
@@ -30,13 +35,23 @@ export class UsersController {
         return this.usersService.findOne(+id);
     }
 
+    @UseGuards(AccessTokenGuard)
     @Patch(':id')
-    update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+    update(
+        @Param('id') id: string,
+        @Body() updateUserDto: UpdateUserDto,
+        @Req() req: Request,
+    ) {
+        if (+id !== req.user?.sub)
+            throw new UnauthorizedException('This is not you');
         return this.usersService.update(+id, updateUserDto);
     }
 
+    @UseGuards(AccessTokenGuard)
     @Delete(':id')
-    remove(@Param('id') id: string) {
+    remove(@Param('id') id: string, @Req() req: Request) {
+        if (+id !== req.user?.sub)
+            throw new UnauthorizedException('This is not you');
         return this.usersService.remove(+id);
     }
 }
